common/types/form.tsのteam関連の型を移動
