+++ features/teams/components/TeamListItem.tsx
@@ -3,8 +3,8 @@ import EditIcon from '@mui/icons-material/Edit';
 import AddIcon from '@mui/icons-material/Add';
 import DeleteIcon from '@mui/icons-material/DeleteForeverOutlined';
 import { FormContext } from '@/context/FormContext';
-import { TeamFormDataParams } from '@/common/types/form';
 import { SessionContext } from '@/context/SessionContext';
+import { TeamFormDataParams } from '../types/context';
 import { InitialInviteTeamFormData } from '../functions/form';
 import { TeamFormContext } from '../contexts/TeamFormContext';
 import { InviteTeamFormContext } from '../contexts/InviteTeamFormContext';
